Trail along with the calculation of the total distance
Make sure to check the comments I provided in the trail script. Change the Color.blue to TrailColor so that it works when you set the color of the trail in the set function. I also provided the getTotal distance function so that you can get the total calculated distance